Force initializing all classes, to avoid missing transitive dependencies
Previously, some of the classes which it detects being loaded, are not actually
initialized. Maybe the Clojure compiler uses the `Class.forName(name,
initialize=false, loader)` variant of `Class.forName` which loads the class but
doesn't initialize it. For example the agent detected the
`com.amazonaws.client.builder.AwsClientBuilder` class as a dependency, but not
that it in turn contains a static field with an
`com.amazonaws.regions.DefaultAwsRegionProviderChain` instance (which in turn
loads a couple more AWS classes and `org.apache.commons.logging`).

The workaround for this would be to either force initialize all classes that
were loaded, or to detect classes which are loaded but not initialized, and
exclude them from the BTI configuration. The former would be an easy loop to
call `Class.forName` for each class. The latter would likely require a bytecode
transformation which inserts a tracing method call to each class's static
initializer to notify when the class is initialized. If `java.lang.Class` would
have an "isInitialized" method, that would make things easier, but I couldn't
find anything like it. Maybe GraalVM would have access to lower level
information that would help here.